pyproject.toml
[project]
name = "ai-agent-lg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "langchain[openai]>=1.0.7",
    "langgraph>=1.0.3",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
